+++ hatano/actions.py
@@ -2,154 +2,103 @@ from hatano.util import Conf
 from copy import deepcopy
 
 import sys
+import os
 
 
 def init(args):
     c = Conf()
     if not c.exists():
-        try:
-            proj_name = input("Project name: ")
-        except KeyboardInterrupt:
-            sys.exit("Aborted")
+        proj = args.name
+        runtime = args.runtime
         conf = {
-                "project": proj_name,
+                "project": proj,
+                "runtime": runtime
                 }
         c.write(conf)
         return
     print("Hatano project exists")
 
 
-def add(args):
+def make(args):
     if args.object == "stage":
-        return add_stage(args)
+        return make_stage(args)
 
     elif args.object == "function":
-        return add_function(args)
+        return make_function(args)
 
 
-def add_stage(args):
-    stage = args.stage
-    copy = args.copy
-    source = args.source
-    bucket = args.bucket
-    runtime = args.runtime
-    domain = args.domain
-    cert = args.cert
+def make_stage(_args, edit=False):
+    args = vars(_args)
+    stage = args["stage"]
 
     with open(f'requirements-{stage}.txt', 'a') as f:
         pass
 
     c = Conf()
     conf = c.read()
-    if stage in conf.get("stages", {}):
+
+    if stage in conf.get("stage", {}) and not edit:
         print(f"Stage {stage} already exists")
         return
 
-    if copy and copy in conf.get("stages", {}):
-        conf["stages"][stage] = deepcopy(conf["stages"][copy])
-        c.write(conf)
+    if stage not in conf.get("stage", {}) and edit:
+        print(f"Stage {stage} doesn't exist")
         return
 
-    if "stages" not in conf:
-        conf["stages"] = {}
+    if "stage" not in conf:
+        conf["stage"] = {}
 
-    conf["stages"][stage] = {
-            "functions": {},
-            "source": source,
-            "runtime": runtime,
-            "bucket": bucket}
-    if domain:
-        conf["stages"][stage]["domain"] = domain
-    if cert:
-        conf["stages"][stage]["cert"] = cert
+    if stage not in conf["stage"]:
+        conf["stage"][stage] = {}
+
+    for key in ["source", "bucket", "domain", "cert"]:
+        if args[key]:
+            conf["stage"][stage][key] = args[key]
 
     c.write(conf)
 
 
-def add_function(args):
-    stage = args.stage
+def make_function(args):
     c = Conf()
     conf = c.read()
-    if stage not in conf.get("stages", {}):
-        add_stage(stage)
-    conf = c.read()
-    if "functions" not in conf["stages"].get(stage, {}):
-        conf["stages"][stage]["functions"] = {}
+    if "function" not in conf:
+        conf["function"] = {}
     func = {
             "handler": args.handler,
             "method": args.method,
             "path": args.path}
-    conf["stages"][stage]["functions"][args.name] = func
+    conf["function"][args.function] = func
     c.write(conf)
 
 
-def remove(args):
-    if args.object == "stage":
-        return rm_stage(args.stage)
+def remove(_args):
+    c = Conf()
+    args = vars(_args)
+    typ = args["object"]
+    name = args[typ]
+    return rm_object(name, typ)
 
-    elif args.object == "function":
-        return rm_function(args)
 
-
-def rm_stage(stage):
+def rm_object(name, typ):
     c = Conf()
     conf = c.read()
-    if "stages" in conf and stage in conf["stages"]:
-        del conf["stages"][stage]
+    if name in conf.get(typ, {}):
+        del conf[typ][name]
     c.write(conf)
-
-
-def rm_function(args):
-    name = args.function
-    stage = args.stage
-    c = Conf()
-    conf = c.read()
-    if "stages" in conf and stage in conf["stages"]:
-        if name in conf["stages"][stage].get("functions", {}):
-            del conf["stages"][stage]["functions"][name]
-    c.write(conf)
+    if typ == "stage":
+        try:
+            os.remove(f"requirements-{name}.txt")
+        except Exception as e:
+            print(e)
 
 def edit(args):
     if args.object == "stage":
-        return edit_stage(args)
+        return make_stage(args, edit=True)
 
     elif args.object == "function":
         return edit_function(args)
 
 
-def edit_stage(args):
-    stage = args.stage
-    copy = args.copy
-    source = args.source
-    bucket = args.bucket
-    runtime = args.runtime
-    domain = args.domain
-    cert = args.cert
-
-    c = Conf()
-    conf = c.read()
-    if stage not in conf.get("stages", {}):
-        print(f"Stage {stage} does not exist")
-        return
-
-    if copy and copy in conf.get("stages", {}):
-        conf["stages"][stage] = deepcopy(conf["stages"][copy])
-        c.write(conf)
-        return
-
-    if source:
-        conf["stages"][stage]["source"] = source
-    if bucket:
-        conf["stages"][stage]["bucket"] = bucket
-    if runtime:
-        conf["stages"][stage]["runtime"] = runtime
-    if domain:
-        conf["stages"][stage]["domain"] = domain
-    if cert:
-        conf["stages"][stage]["cert"] = cert
-
-
-    c.write(conf)
 
 
 def edit_function(args):
@@ -157,20 +106,20 @@ def edit_function(args):
     c = Conf()
     conf = c.read()
 
-    if stage not in conf.get("stages", {}):
+    if stage not in conf.get("stage", {}):
         print(f"Stage {stage} does not exist")
         return
 
     conf = c.read()
-    if "functions" not in conf["stages"].get(stage, {}):
+    if "function" not in conf["stage"].get(stage, {}):
         print(f"No functions defined in stage {stage}")
         return
 
-    if args.name not in conf["stages"][stage]["functions"]:
+    if args.name not in conf["stage"][stage]["function"]:
         print(f"Function {args.name} not defined in stage {stage}")
         return
 
-    func = conf["stages"][stage]["functions"][args.name]
+    func = conf["stage"][stage]["function"][args.name]
     if args.handler:
         func["handler"] = args.handler
     if args.method:
@@ -178,7 +127,7 @@ def edit_function(args):
     if args.path:
         func["path"] = args.path
 
-    conf["stages"][stage]["functions"][args.name] = func
+    conf["stage"][stage]["function"][args.name] = func
     c.write(conf)