+++ hatano/iam.py
@@ -134,7 +134,8 @@ class IamRole:
         self.stage = stage
         self.name = fnargs.get("name")
         c = Conf()
-        self.project, stg_conf = c.get_stage(stage)
+        conf = c.read()
+        self.project = conf["project"]
         self.iam = boto3.client('iam')
         self.fullname = f"{self.project}-{self.name}-{self.stage}"
 
@@ -142,7 +143,8 @@ class IamRole:
         role = self.iam.create_role(
                 Path=f"/{self.project}/",
                 RoleName=self.fullname,
-                AssumeRolePolicyDocument=json.dumps(trust)
+                AssumeRolePolicyDocument=json.dumps(trust),
+                Description="Created automatically by Hatano"
                 )
         return role