+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.3 (TBA)
+
+- Threaded function deployment
+- Conductor replaces deploy.py and update.py
+- Clean removes requirements-{stage}.txt file
+- Deploy by function, better support for multi-stage projects
+
 ## 1.0.2 (20190621)
 
 S3 integration